add line //return storeItem object
Maybe this was implied but I don't think so.  Other functions in the assignment explicitly say things like 'return the users array' after the required changes are completed.  I had this correct but kept making changes for almost an hour until I saw on the slack channel that you had to return the object after you made the changes